% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Functions_and_Methods.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomial Choose Main Function}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials - numeric}

\item{k}{number of successes - numeric}
}
\value{
a number of combinations - numeric
}
\description{
calculates the number of combinations in which k successes can occur in n trials
}
\examples{
bin_choose(n = 5, k = 2)
bin_choose(5, 0)
bin_choose(5, 1:3)
}
