% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Functions_and_Methods.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial Cumulative Main Function}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials - numeric}

\item{prob}{probability of each trial (more than 0 and less than 1)}
}
\value{
a data.frame with two classes: c("bincum", "data.frame")
}
\description{
computes a data frame with the probability distribution along with cumulative probability: sucesses in the first column, prob in the second
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
