% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Functions_and_Methods.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial Kurtosis Function}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials - numeric}

\item{prob}{probability of each trial (more than 0 and less than 1)}
}
\value{
the kurtosis of binomial distribution
}
\description{
computes the kurtosis of binomial distribution
}
\examples{
bin_kurtosis(10, 0.3)
}
