% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Functions_and_Methods.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial Mean Function}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials - numeric}

\item{prob}{probability of each trial (more than 0 and less than 1)}
}
\value{
the mean of binomial distribution
}
\description{
computes the mean of binomial distribution
}
\examples{
bin_mean(10, 0.3)
}
