% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Functions_and_Methods.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial Mode Function}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials - numeric}

\item{prob}{probability of each trial (more than 0 and less than 1)}
}
\value{
the mode of binomial distribution
}
\description{
computes the mode of binomial distribution
}
\examples{
bin_mode(10, 0.3)
}
