% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Functions_and_Methods.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial Probability Main Function}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{number of successes - numeric}

\item{trials}{number of trials - numeric}

\item{prob}{probability of each trial (more than 0 and less than 1)}
}
\value{
a probability of the binomial
}
\description{
compute the binomial probabilities based on the success, independent trials and probability
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)
bin_probability(success = 0:2, trials = 5, prob = 0.5)
bin_probability(success = 55, trials = 100, prob = 0.45)
}
