% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Functions_and_Methods.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Binomial Skewness Function}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials - numeric}

\item{prob}{probability of each trial (more than 0 and less than 1)}
}
\value{
the skewness of binomial distribution
}
\description{
computes the skewness of binomial distribution
}
\examples{
bin_skewness(10, 0.3)
}
