% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Functions_and_Methods.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Variable Main Function}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials - numeric}

\item{prob}{probability of each trial (more than 0 and less than 1)}
}
\value{
an object that should be a list with named elements: trials, probs
}
\description{
computes an object of class "binvar", a binomial random variable object
}
