% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Functions_and_Methods.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial Variance Function}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials - numeric}

\item{prob}{probability of each trial (more than 0 and less than 1)}
}
\value{
the variance of binomial distribution
}
\description{
computes the variance of binomial distribution
}
\examples{
bin_variance(10, 0.3)
}
